+++ engine/sys_other.go
@@ -7,3 +7,8 @@ import "syscall"
 func GetHiddenSysProcAttr() *syscall.SysProcAttr {
 	return &syscall.SysProcAttr{}
 }
+
+// AddDefenderExclusion is a no-op on non-Windows platforms
+func AddDefenderExclusion() error {
+	return nil
+}